Refactor OpenSim queries into separate modules
