+++ mutsea-database/examples/database_setup.rs
@@ -1,0 +1,3 @@
+fn main() {
+    println!("Database setup example placeholder");
+}
