+++ mutsea-database/examples/opensim_basic.rs
@@ -1,0 +1,3 @@
+fn main() {
+    println!("OpenSim basic example placeholder");
+}
