+++ mutsea-database/src/opensim/queries/mod.rs
@@ -1,0 +1,6 @@
+// src/opensim/queries/mod.rs
+//! OpenSim database query modules
+
+pub mod user_queries;
+pub mod asset_queries;
+pub mod region_queries;
